fix: disable unconformity with the same name as the feature when isosurfacing that feature.
for example when adding an unconformity to the bottom of a unit at value 0, we want to make sure this scalar field is never < 0 but if we do this we can't isosurface 0 because the cubes containing 0 isovalue have half the nodes being nan. To avoid this, when using the feature.surfaces() method, the region which is related to this feature can be disabled.